% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaGMM.R
\name{collect_replicates}
\alias{collect_replicates}
\title{Pull protein's chromatograms across replicates into matrix}
\usage{
collect_replicates(protein, replicates)
}
\arguments{
\item{protein}{The protein name as it appears in the rowname of \code{replicates}}

\item{replicates}{A list of matrices representing replicates with protein
names as rownames and fractions as columns.}
}
\value{
A matrix with the each replicate of the protein in rows
}
\description{
Pull protein's chromatograms across replicates into matrix
}
