\name{DeltaGMM-package}
\alias{DeltaGMM-package}
\alias{DeltaGMM}
\docType{package}
\title{
\packageTitle{DeltaGMM}
}
\description{
\packageDescription{DeltaGMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DeltaGMM}
\packageIndices{DeltaGMM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{DeltaGMM}

Maintainer: \packageMaintainer{DeltaGMM}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
# simple examples of the most important functions
}
