% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaGMM.R
\name{deltaGMM}
\alias{deltaGMM}
\title{Scores differential protein interactions in CF-MS data}
\usage{
deltaGMM(
  condition1,
  condition2,
  score = c("fstat", "manhattan"),
  imputectrl = setimputectrl(),
  gmmctrl = setgmmctrl(),
  ncores = TRUE
)
}
\arguments{
\item{condition1, condition2}{A pair of matrices or a paired list of matrices
representing replicates of chromatograms. Proteins are represented as rows
and fractions are represented as columns. Replicates must have matching
number of fractionations. Only proteins appearing in both conditions will be
used.}

\item{score}{The scoring metric to use. Defaults to \code{"fstat"}.}

\item{imputectrl}{An optional named list of class "imputectrl" of parameters
for imputation. See \code{\link[=setimputectrl]{setimputectrl()}} for names of settable control values and
their effects.}

\item{gmmctrl}{An optional named list of class "gmmctrl" of parameters
for imputation. See \code{\link[=setgmmctrl]{setgmmctrl()}} for names of settable control values and
their effects.}

\item{ncores}{Number of cores to use for parallel computation. Default of 1
specifies serial computation}
}
\value{
A named list of proteins with scores.
}
\description{
Identifies changes in a protein's complex membership between two conditions
by fitting a Gaussian Mixture Model to the protein chromatogram in each
condition as a null model where a GMM is fit to both conditions
simulatenously. Proteins are then ranked by some metric (such as the
F-statistic) to identify differential protein interactions.
}
