% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_chromatograms.R
\name{distance_from_quantified}
\alias{distance_from_quantified}
\title{Find distance to nearest quantified fraction}
\usage{
distance_from_quantified(chrom)
}
\arguments{
\item{chrom}{A numeric vector of a protein chromatogram}
}
\value{
A numeric vector with same length as chrom with distance from
nearest quantified fraction for each fraction
}
\description{
Find distance to nearest quantified fraction
}
