% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_protein.R
\name{fit_protein}
\alias{fit_protein}
\title{Fits a Gaussian Mixture model to protein chromatograms}
\usage{
fit_protein(chromatograms, imputectrl = imputectrl(), gmmctrl = gmmctrl())
}
\arguments{
\item{chromatograms}{A matrix of raw chromatograms in rows (replicates)
and fractions in columns to fit gaussians to. A single numeric vector will be
coerced to a matrix with one row.}

\item{imputectrl}{An optional named list of class "imputectrl" of parameters
for imputation. See \code{\link[=imputectrl]{imputectrl()}} for names of settable control values and
their effects.}

\item{gmmctrl}{An optional named list of class "gmmctrl" of parameters
for imputation. See \code{\link[=gmmctrl]{gmmctrl()}} for names of settable control values and
their effects.}
}
\value{
A named list containing the fitted model for the protein and other
details
}
\description{
Fits one GMM to the raw chromatograms in each condition and one GMM too all
the chromatograms as if they were from the same condition to make a null
model. A specified score is then calculated from features of the fits to rank
the amount of differential change
}
