% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_weights.R
\name{impute_chromatograms}
\alias{impute_chromatograms}
\title{Impute a matrix of chromatograms based on provided parameters}
\usage{
impute_chromatograms(chromatograms, imputectrl = imputectrl())
}
\arguments{
\item{chromatograms}{A matrix of raw chromatograms in rows (replicates)
and fractions in columns to fit gaussians to. A single numeric vector will be
coerced to a matrix with one row.}

\item{imputectrl}{An optional named list of class "imputectrl" of parameters
for imputation. See \code{\link[=imputectrl]{imputectrl()}} for names of settable control values and
their effects.}
}
\value{
a matrix of chromatograms with imputed missing values
}
\description{
Impute a matrix of chromatograms based on provided parameters
}
