% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gmm.R
\name{predict_gmm}
\alias{predict_gmm}
\title{Returns the fitted GMM model from the output of \code{\link[=fit_protein]{fit_protein()}}}
\usage{
predict_gmm(model)
}
\arguments{
\item{model}{a object output from \code{\link[=fit_n_gaussians]{fit_n_gaussians()}} containing \code{coefs},
\code{n_gaussians}, and \code{fractions} from the NLS model}
}
\value{
a vector of the fitted GMM
}
\description{
Returns the fitted GMM model from the output of \code{\link[=fit_protein]{fit_protein()}}
}
