% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{setgmmctrl}
\alias{setgmmctrl}
\title{Parameters used to control GMM fitting}
\usage{
setgmmctrl(
  max_gaussians = 5,
  criterion_fn = AICc,
  max_iterations = 5,
  init_method = c("guess", "random"),
  height_min = 0.1,
  height_max = 1.1,
  variance_min = 2,
  variance_max = 0,
  rssweights = TRUE,
  return_fit = FALSE
)
}
\arguments{
\item{max_gaussians}{The maximum number of Gaussians to fit. Defaults to 5.}

\item{criterion_fn}{function used in model selection.
Choose one of \code{\link[=AICc]{AICc()}} (default), \code{\link[=AIC]{AIC()}}, \code{\link[=BIC]{BIC()}}.}

\item{max_iterations}{The number of times to try fitting the curve with
different initial conditions.}

\item{init_method}{The method used to select the initial conditions for
nonlinear least squares optimization. One of \code{"guess"} (default) or \code{ "random"}.}

\item{height_min}{Ratio to the max value within the chromatograms for which
gaussians with lower heights will be filtered out. Defaults to 0.15.}

\item{height_max}{Ratio to the max value within the chromatograms for
which gaussians with larger heights will be filtered out. Defaults to 1.5.}

\item{variance_min}{Minimum variance for fitted gaussians. Defaults to 1.}

\item{variance_max}{Maximum variance for fitted gaussians. \code{0} sets maximum
variance to half the fractions (default).}

\item{rssweights}{Calculate RSS with imputed values adjusted by their weight.}

\item{return_fit}{Returns the fitted model from \code{\link[=nls]{nls()}}. Useful for
exporatory analysis. Defaults to \code{FALSE}.}
}
\value{
An named list of class \code{gmmctrl} with specified parameters
}
\description{
Parameters used to control GMM fitting
}
