% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{setimputectrl}
\alias{setimputectrl}
\title{Parameters used to control imputation of chromatograms}
\usage{
setimputectrl(
  imputation = c("dynamic", "fill_gaps", "weights", "none"),
  max_dist = 3,
  na_weight = 0.1,
  log_k = 1.5,
  log_x = 4
)
}
\arguments{
\item{imputation}{Type of imputation to use for missing values.
\itemize{
\item \code{"dynamic"} (the default): Imputes NAs as 0, then dynamically sets the
weight based on the distance to the nearest quantified fraction with
\code{\link[=distance_from_quantified]{distance_from_quantified()}}. Weights are passed to \code{\link[=nls]{nls()}}.
\item \code{"weights"}: Used with \code{na_weights} argument. Sets all missing values to 0
and sets a fixed weight based on \code{na_weights}.
\item \code{"fill_gaps"}: Used with \code{max_dist} argument. Sets missing values that are
greater or equal to\code{max_dist} from the nearest quantified fraction to 0 and
leaves closer missing values as NA. This was the first method of imputation
developed to make sure nls doesn't fit gaussians to large gaps without
penalization.
\item \code{"none"}: No imputation and raw values are passed to \code{\link[=nls]{nls()}}.
}}

\item{max_dist}{Maximum gap of unquantified fractions before imputation}

\item{na_weight}{The weight of missing values imputed as 0}

\item{log_k, log_x}{The \code{k} and \code{x_0} value of the logistic function used in
dynamic imputation of weights. \code{k} represents the value of the sigmoid
midpoint or how many fractions away from the nearest quantified fraction
for the imputed weights to be 0.5. \code{k} is the steepness of curve transition
from 0 to 1. Default of 1.5 and 4 were determined emperically.}
}
\value{
An named list of class \code{imputectrl} with specified parameters
}
\description{
Parameters used to control imputation of chromatograms
}
