% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_params.R
\name{smooth_chromatogram}
\alias{smooth_chromatogram}
\title{Smooth chromatogram with moving average and option to impute NAs}
\usage{
smooth_chromatogram(chromatogram, impute_NA = FALSE, smooth_width = 4)
}
\value{
Smoothed chromatogram
}
\description{
Smooth chromatogram with moving average and option to impute NAs
}
