% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_params.R
\name{start_params}
\alias{start_params}
\title{Set starting parameters for NLS curve fitting}
\usage{
start_params(
  chromatogram,
  n_gaussians,
  gmmctrl = setgmmctrl(),
  sigma_default = 2
)
}
\arguments{
\item{chromatogram}{A numeric vector of a smoothed protein chromatogram}

\item{n_gaussians}{The number of gaussian parameters to output}

\item{gmmctrl}{An optional named list of class "gmmctrl" of parameters
for imputation. See \code{\link[=setgmmctrl]{setgmmctrl()}} for names of settable control values and
their effects.}

\item{sigma_default}{The default mean initial value of sigma}
}
\value{
A named list containing \code{A}, \code{mu}, \code{sigma} each with length of
\code{n_gaussians}
}
\description{
Construct a set of initial conditions for curve fitting using nonlinear least
squares using a mixture of Gaussians. This method finds local maxima within
the chromatogram, orders them by their magnitude from local maxima and uses
the positions and heights of these local maxima (+/- some random noise) as
initial conditions. The "random" method simply picks random values within
the fraction and intensity intervals as starting points for Gaussian
curve-fitting. The initial value of sigma is set by default to a random
number within +/- 0.5 of two for both modes; this is based on our manual
inspection of a large number of chromatograms.
Derived from PrInCE's \code{make_initial_conditions}
}
