% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit_n_gaussians.R
\name{test_fit_n_gaussians}
\alias{test_fit_n_gaussians}
\title{Fits a Gaussian Mixture Model to chromatograms}
\usage{
test_fit_n_gaussians(
  chromatograms,
  n_gaussians,
  imputectrl = setimputectrl(),
  gmmctrl = setgmmctrl()
)
}
\arguments{
\item{chromatograms}{A matrix of raw chromatograms in rows (replicates)
and fractions in columns to fit gaussians to. A single numeric vector will be
coerced to a matrix with one row.}

\item{n_gaussians}{The number of gaussian components to fit.}

\item{imputectrl}{An optional named list of class "imputectrl" of parameters
for imputation. See \code{\link[=setimputectrl]{setimputectrl()}} for names of settable control values and
their effects.}

\item{gmmctrl}{An optional named list of class "gmmctrl" of parameters
for imputation. See \code{\link[=setgmmctrl]{setgmmctrl()}} for names of settable control values and
their effects.}
}
\value{
A fitted model of the chromatograms with various details
}
\description{
Attempts to fit a Gaussian Mixture Model to the provided chromatograms with
\code{n_gaussians} number of components using \code{\link[=nls]{nls()}}. Performs imputation of
chromatograms based on \code{imputectrl} parameters.
}
