% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_protein.R
\name{fit_protein}
\alias{fit_protein}
\title{Fits a Gaussian Mixture model to protein chromatograms}
\usage{
fit_protein(
  chromatograms,
  imputectrl = setimputectrl(),
  gmmctrl = setgmmctrl()
)
}
\arguments{
\item{chromatograms}{A matrix of raw chromatograms in rows (replicates)
and fractions in columns to fit gaussians to. A single numeric vector will be
coerced to a matrix with one row.}

\item{imputectrl}{An optional named list of class "imputectrl" of parameters
for imputation. See \code{\link[=setimputectrl]{setimputectrl()}} for names of settable control values and
their effects.}

\item{gmmctrl}{An optional named list of class "gmmctrl" of parameters
for imputation. See \code{\link[=setgmmctrl]{setgmmctrl()}} for names of settable control values and
their effects.}
}
\value{
A named list containing the fitted model for the protein and other
details
}
\description{
Attempts to fit multiple GMM to the provided chromatograms with different
number of gaussian components from 1 to \code{max_gaussians}. Selects best model
based on the specified information criterion in \code{gmmctrl}.
}
